create feedback if already checkin in my booking detail

+++ fhotel_1/lib/data/models/feedback.dart
@@ -1,7 +1,7 @@
 class Feedbacks {
   String? feedbackId;
   String? reservationId;
-  double? hotelRating;
+  int? hotelRating;
   String? comment;
   String? createdDate;
